backend/requirements.txt
fastapi[standard]
sqlmodel
sqlalchemy
psycopg2-binary
pytest
requests
httpx
pydantic-settings
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
python-multipart
